src/backend/pyproject.toml
[project]
name = "slidekick"
version = "0.1.0"
description = "Slidekick - Backend API"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Nick Muoh", email = "nickspirit3@gmail.com" },
    { name = "Snehal Vartak", email = "snehal.inusa@gmail.com" },
]

dependencies = [
    "fastapi>=0.115.0",
    "google-genai>=1.55.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.1",
]

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[project.scripts]
slidekick = "slidekick.slide_deck_client:main"

[tool.ruff]
line-length = 90


[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"